pyproject.toml
[project]
name = "ELVIS-C"
version = "0.0.1"
description = "Minimal interactive environment for ELVIS-Video with interventions"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "gymnasium>=0.29",
  "numpy>=1.24",
  "pillow>=10.0",
  "imageio>=2.34",
  "tqdm>=4.66",
  "pydantic>=2.8"
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["elvis_env*"]